containers: fix a crash when --clean is executed twice in a row

+++ containers/scripts/deploy.sh
@@ -97,7 +97,7 @@ clean() {
     images=$(podman images -q 'osh-*' | paste -s -d' ')
     # podman images -q has a defined format
     # shellcheck disable=2086
-    podman rmi -f $images
+    [ -z "$images" ] || podman rmi -f $images
 }
 
 while [[ $# -gt 0 ]]; do
